Upgrade to the new Jinja2 compiler syntax  (#51)
I tracked a segfault on Linux to a bug in GMT that crashed when the 
path names were too long (like the conda placehold_... prefixes). 
This is fixed in PR #52 for the devel branch and will be backported to 
5.4.5 which should come out soon. For now, I tested the built packaged 
on the Docker container locally and it works fine. I'll comment out the 
tests for now to get this built and will put them back with the new release.