Do not pack GSHHG and DCW data for Linux and macOS
In GMT 6.1.0, the default value of COPY_GSHHG and COPY_DCW changes from
FALSE to TRUE, but for conda-forge package, we don't want to copy DCW and
GSHHG data.

The only exception is Windows. For unknown reason, we have to copy
GSHHG and DCW data to make the Windows build pass.