Do not pack GSHHG and DCW data for Linux and macOS (#98)
GMT needs the GSHHG and DCW data to work. conda-forge provides separate packages
for gshhg and dcw, so this feedstock doesn't need to package the GSHHG and DCW data
into the gmt package. One exception is Windows. For unknown reasons,
we have to copy GSHHG and DCW data in the gmt package to make the
Windows build pass.

In GMT 6.1.0, the default values of COPY_GSHHG and COPY_DCW changes from FALSE
to TRUE, so we have to set the two variables to FALSE on Linux and macOS
to not pack the data.